# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "이웃집멍냥 백엔드 API 문서입니다.",
        "title": "이웃집멍냥 API 문서",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "petsnextdoordev@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.7.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "FirebaseAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "사용자 목록을 조회합니다.",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "페이지 번호",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "페이지 사이즈",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "닉네임 (완전 일치)",
                        "name": "nickname",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserWithoutPrivateInfoList"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "파이어베이스 가입 이후 정보를 입력 받아 유저를 생성합니다.",
                "parameters": [
                    {
                        "description": "사용자 회원가입 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.RegisterUserView"
                        }
                    }
                }
            }
        },
        "/users/check/nickname": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "닉네임 중복 여부를 조회합니다.",
                "parameters": [
                    {
                        "description": "사용자 닉네임 중복 조회 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CheckNicknameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.CheckNicknameView"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "security": [
                    {
                        "FirebaseAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "내 프로필 정보를 조회합니다.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.MyProfileView"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "FirebaseAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "내 프로필 정보를 수정합니다.",
                "parameters": [
                    {
                        "description": "사용자 프로필 수정 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UpdateUserView"
                        }
                    }
                }
            }
        },
        "/users/me/pets": {
            "get": {
                "security": [
                    {
                        "FirebaseAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users",
                    "pets"
                ],
                "summary": "내 반려동물 목록을 조회합니다.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pet.FindMyPetsView"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "FirebaseAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users",
                    "pets"
                ],
                "summary": "내 반려동물을 등록합니다.",
                "parameters": [
                    {
                        "description": "반려동물 등록 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pet.AddPetsToOwnerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/status": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "이메일로 유저의 가입 상태를 조회합니다.",
                "parameters": [
                    {
                        "description": "사용자 가입 상태 조회 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserStatusView"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "pet.AddPetRequest": {
            "type": "object",
            "required": [
                "birth_date",
                "breed",
                "name",
                "neutered",
                "pet_type",
                "sex",
                "weight_in_kg"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "neutered": {
                    "type": "boolean"
                },
                "pet_type": {
                    "enum": [
                        "dog",
                        "cat"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/pet.PetType"
                        }
                    ]
                },
                "sex": {
                    "enum": [
                        "male",
                        "female"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/pet.PetSex"
                        }
                    ]
                },
                "weight_in_kg": {
                    "type": "number"
                }
            }
        },
        "pet.AddPetsToOwnerRequest": {
            "type": "object",
            "required": [
                "pets"
            ],
            "properties": {
                "pets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pet.AddPetRequest"
                    }
                }
            }
        },
...
        "pet.FindMyPetsView": {
            "type": "object",
            "properties": {
                "pets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pet.PetView"
                    }
                }
            }
        },
        "pet.PetSex": {
            "type": "string",
            "enum": [
                "male",
                "female"
            ],
            "x-enum-varnames": [
                "PetSexMale",
                "PetSexFemale"
            ]
        },
        "pet.PetType": {
            "type": "string",
            "enum": [
                "dog",
                "cat"
            ],
            "x-enum-varnames": [
                "PetTypeDog",
                "PetTypeCat"
            ]
        },
        "pet.PetView": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "neutered": {
                    "type": "boolean"
                },
                "pet_type": {
                    "$ref": "#/definitions/pet.PetType"
                },
                "sex": {
                    "$ref": "#/definitions/pet.PetSex"
                },
                "weight_in_kg": {
                    "type": "number"
                }
            }
        },
...
        "user.CheckNicknameRequest": {
            "type": "object",
            "required": [
                "nickname"
            ],
            "properties": {
                "nickname": {
                    "type": "string"
                }
            }
        },
        "user.CheckNicknameView": {
            "type": "object",
            "properties": {
                "isAvailable": {
                    "type": "boolean"
                }
            }
        },
        "user.FirebaseProviderType": {
            "type": "string",
            "enum": [
                "email",
                "google",
                "apple",
                "kakao"
            ],
            "x-enum-varnames": [
                "FirebaseProviderTypeEmail",
                "FirebaseProviderTypeGoogle",
                "FirebaseProviderTypeApple",
                "FirebaseProviderTypeKakao"
            ]
        },
        "user.MyProfileView": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fbProviderType": {
                    "$ref": "#/definitions/user.FirebaseProviderType"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "profileImageUrl": {
                    "type": "string"
                }
            }
        },
        "user.RegisterUserRequest": {
            "type": "object",
            "required": [
                "email",
                "fbProviderType",
                "fbUid",
                "fullname",
                "nickname"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "fbProviderType": {
                    "$ref": "#/definitions/user.FirebaseProviderType"
                },
                "fbUid": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "profileImageId": {
                    "type": "integer"
                }
            }
        },
        "user.RegisterUserView": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fbProviderType": {
                    "$ref": "#/definitions/user.FirebaseProviderType"
                },
                "fbUid": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "profileImageUrl": {
                    "type": "string"
                }
            }
        },
        "user.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "profileImageId": {
                    "type": "integer"
                }
            }
        },
        "user.UpdateUserView": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fbProviderType": {
                    "$ref": "#/definitions/user.FirebaseProviderType"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "profileImageUrl": {
                    "type": "string"
                }
            }
        },
        "user.UserRegistrationStatus": {
            "type": "string",
            "enum": [
                "NOT_REGISTERED",
                "REGISTERED"
            ],
            "x-enum-varnames": [
                "UserStatusNotRegistered",
                "UserStatusRegistered"
            ]
        },
        "user.UserStatusRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "user.UserStatusView": {
            "type": "object",
            "properties": {
                "fbProviderType": {
                    "$ref": "#/definitions/user.FirebaseProviderType"
                },
                "status": {
                    "$ref": "#/definitions/user.UserRegistrationStatus"
                }
            }
        },
        "user.UserWithoutPrivateInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "profileImageUrl": {
                    "type": "string"
                }
            }
        },
        "user.UserWithoutPrivateInfoList": {
            "type": "object",
            "properties": {
                "is_last_page": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.UserWithoutPrivateInfo"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "FirebaseAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}